+++ Juego/src/Enemigo.cpp
@@ -36,7 +36,7 @@ Sala* Enemigo::getSala()
     return estoy;
 }
 
-void Enemigo::setPosiciones(int nx,int ny,int nz)
+void Enemigo::setPosiciones(float nx,float ny,float nz)
 {
     x = nx;
     y = ny;
