+++ Juego/src/INdrawable.hpp
@@ -9,7 +9,7 @@ class INdrawable
 {
 
 public:
-    virtual void setPosiciones(int nx,int ny,int nz)=0;//modifica las posiciones de la figura
+    virtual void setPosiciones(float nx,float ny,float nz)=0;//modifica las posiciones de la figura
     virtual void setID(int)=0;//nos sirve para definir el id que tiene esta figura
     virtual int getID()=0;//nos sirve para saber que id tiene esta figura
